Update URL to handle requests with more than one word
Example of situation when it's needed :

Say I want to have an article related to "Féminiser sa voix".
If I ask "Féminiser", it does not return the right article.
If I ask "voix", it does not return the right article either.
I need to be able to ask for "Féminiser sa voix".