Fix an HTTP Parameter Pollution bug
Before this commit, if a user put a char `&` in keywords (which should never happen), due to the concatenation without encoding the value, the server will handle only what comes before this char as the parameter value. What comes after will be considered as another HTTP parameter.

Example:
If a user is searching `test & foobar`,
the URL will be as `term=test & foobar`.
Thus, the value of `term` parameter will be `test ` and the following parameter is ` foobar`.
If we URL-encode the searched terms, the URL will be `term=test%20%26%20foobar`.
Thus, the value of `term` parameter will be `test & foobar`, as terms searched by the user.

Usually an HPP is a vulnerability, however, here it's not a vulnerability, only a bug.